+++ torchtune/_cli/subcommand.py
@@ -4,10 +4,8 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-from typing import Protocol
 
-
-class Subcommand(Protocol):
+class Subcommand:
     def __init__(self, *args, **kwargs):
         pass
