+++ torchtune/utils/argparse.py
@@ -11,9 +11,9 @@ from typing import List, Tuple
 from omegaconf import OmegaConf
 
 
-class TuneArgumentParser(argparse.ArgumentParser):
+class TuneRecipeArgumentParser(argparse.ArgumentParser):
     """
-    TuneArgumentParser is a helpful utility subclass of the argparse ArgumentParser that
+    TuneRecipeArgParser is a helpful utility subclass of the argparse ArgumentParser that
     adds a builtin argument "config". The config argument takes a file path to a yaml file
     and will load in argument defaults from the yaml file. The yaml file must only contain
     argument names and their values and nothing more, it does not have to include all of the
